requirements.txt
djangorestframework
django-cors-headers
requests
django-sslserver
django-ace
git+https://github.com/RANUX/django-ace.git@main#egg=django-ace